local-assistant-pro/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "local-assistant-pro"
version = "0.1.0"
description = "Offline local AI assistant with RAG, STT/TTS, screen capture, skills"
authors = [{name = "You"}]
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.111",
  "uvicorn[standard]>=0.30",
  "jinja2>=3.1",
  "pydantic>=2.7",
  "python-dotenv>=1.0",
  "PyYAML>=6.0",
  "loguru>=0.7",
  "watchdog>=4.0",
  "llama-cpp-python==0.2.90",
  "chromadb==0.5.3",
  "langchain==0.2.17",
  "langchain-community==0.2.10",
  "sentence-transformers>=3.0",
  "mss>=9.0",
  "pillow>=10.3",
  "imageio>=2.34"
]

[project.optional-dependencies]
speech = [
  "faster-whisper>=1.0; platform_system!='Windows' or platform_machine!='arm64'",
  "torch>=2.1; extra == 'speech'",
]

[project.scripts]
assistant = "local_assistant_pro.cli:app"
